i18n: Created ThemePicker so it can be translated in both tabs of the share modal
