+++ public/locales/pseudo-LOCALE/messages.po
@@ -527,22 +527,6 @@ msgid "share-modal.link.shorten-url"
 msgstr ""
 
 #: public/app/features/dashboard/components/ShareModal/ShareLink.tsx
-msgid "share-modal.link.theme"
-msgstr ""
-
-#: public/app/features/dashboard/components/ShareModal/ShareLink.tsx
-msgid "share-modal.link.theme-current"
-msgstr ""
-
-#: public/app/features/dashboard/components/ShareModal/ShareLink.tsx
-msgid "share-modal.link.theme-dark"
-msgstr ""
-
-#: public/app/features/dashboard/components/ShareModal/ShareLink.tsx
-msgid "share-modal.link.theme-light"
-msgstr ""
-
-#: public/app/features/dashboard/components/ShareModal/ShareLink.tsx
 msgid "share-modal.link.time-range-description"
 msgstr ""
 
@@ -612,6 +596,22 @@ msgstr ""
 
 #: public/app/features/dashboard/components/ShareModal/ShareSnapshot.tsx
 msgid "share-modal.snapshot.timeout-description"
+msgstr ""
+
+#: public/app/features/dashboard/components/ShareModal/ThemePicker.tsx
+msgid "share-modal.theme-picker.current"
+msgstr ""
+
+#: public/app/features/dashboard/components/ShareModal/ThemePicker.tsx
+msgid "share-modal.theme-picker.dark"
+msgstr ""
+
+#: public/app/features/dashboard/components/ShareModal/ThemePicker.tsx
+msgid "share-modal.theme-picker.field-name"
+msgstr ""
+
+#: public/app/features/dashboard/components/ShareModal/ThemePicker.tsx
+msgid "share-modal.theme-picker.light"
 msgstr ""
 
 #: public/app/core/components/SharedPreferences/SharedPreferences.tsx
